+++ Sources/Shared/Publication/Metadata.swift
@@ -31,6 +31,8 @@ public struct Metadata: Hashable, Loggable, WarningLogger {
     public let modified: Date?
     public let published: Date?
     public let languages: [String]  // BCP 47 tag
+    // Main language of the publication.
+    public let language: Language?
     public let sortAs: String?
     public let subjects: [Subject]
     public let authors: [Contributor]
@@ -103,6 +105,7 @@ public struct Metadata: Hashable, Loggable, WarningLogger {
         self.modified = modified
         self.published = published
         self.languages = languages
+        self.language = languages.first.map { Language(code: .bcp47($0)) }
         self.sortAs = sortAs
         self.subjects = subjects
         self.authors = authors
@@ -151,6 +154,7 @@ public struct Metadata: Hashable, Loggable, WarningLogger {
         self.modified = parseDate(json.pop("modified"))
         self.published = parseDate(json.pop("published"))
         self.languages = parseArray(json.pop("language"), allowingSingle: true)
+        self.language = languages.first.map { Language(code: .bcp47($0)) }
         self.sortAs = json.pop("sortAs") as? String
         self.subjects = [Subject](json: json.pop("subject"), warnings: warnings)
         self.authors = [Contributor](json: json.pop("author"), warnings: warnings,  normalizeHREF: normalizeHREF)