+++ Sources/Shared/Toolkit/Cancellable.swift
@@ -14,10 +14,21 @@ public protocol Cancellable {
 
 /// A `Cancellable` object saving its cancelled state.
 public final class CancellableObject: Cancellable {
+
     public private(set) var isCancelled = false
+    private let onCancel: () -> Void
+
+    public init(onCancel: @escaping () -> Void = {}) {
+        self.onCancel = onCancel
+    }
 
     public func cancel() {
+        guard !isCancelled else {
+            return
+        }
+
         isCancelled = true
+        onCancel()
     }
 }
 
@@ -57,5 +68,17 @@ public final class MediatorCancellable: Cancellable {
         isCancelled = true
         cancellable?.cancel()
         cancellable = nil
+    }
+}
+
+public extension Cancellable {
+    /// Convenience to mediate a cancellable in a call chain.
+    ///
+    /// ```
+    /// apiReturningACancellable()
+    ///     .mediate(by: mediator)
+    /// ```
+    func mediated(by mediator: MediatorCancellable) {
+        mediator.mediate(self)
     }
 }