+++ Sources/Shared/Toolkit/Extensions/Result.swift
@@ -17,6 +17,10 @@ extension Result {
         return try? get()
     }
 
+    func get(or def: Success) -> Success {
+        (try? get()) ?? def
+    }
+
     func `catch`(_ recover: (Failure) -> Self) -> Self {
         if case .failure(let error) = self {
             return recover(error)
@@ -24,4 +28,19 @@ extension Result {
         return self
     }
 
+    func eraseToAnyError() -> Result<Success, Error> {
+        mapError { $0 as Error }
+    }
+}
+
+extension Result where Failure == Error {
+    func tryMap<T>(_ transform:(Success) throws -> T)  -> Result<T, Error> {
+        flatMap {
+            do {
+                return .success(try transform($0))
+            } catch {
+                return .failure(error)
+            }
+        }
+    }
 }